+++ criu/include/parasite-syscall.h
@@ -23,7 +23,6 @@ struct rt_sigframe;
 struct parasite_ctl;
 struct parasite_thread_ctl;
 
-extern int parasite_dump_sigacts_seized(struct parasite_ctl *ctl, struct pstree_item *);
 extern int parasite_dump_itimers_seized(struct parasite_ctl *ctl, struct pstree_item *);
 
 struct proc_posix_timers_stat;
